+++ test_project/test_gomock/user/user.go
@@ -10,6 +10,6 @@ func NewUser(p person.Male) *User {
 	return &User{Person: p}
 }
 
-func (u *User) GetUserInfo(id int64) int64 {
+func (u *User) GetUserInfo(id int64) int {
 	return u.Person.Get(id)
 }
